+++ services/billing-core/src/main/java/br/com/arinelli/pay/billing/clients/DocumentType.java
@@ -1,0 +1,6 @@
+package br.com.arinelli.pay.billing.clients;
+
+public enum DocumentType {
+    CPF,
+    CNPJ
+}