+++ services/billing-core/src/main/java/br/com/arinelli/pay/billing/clients/DuplicateDocumentException.java
@@ -1,0 +1,8 @@
+package br.com.arinelli.pay.billing.clients;
+
+public class DuplicateDocumentException extends RuntimeException {
+
+    public DuplicateDocumentException(String document) {
+        super("Documento já cadastrado: " + document);
+    }
+}